Convert memory.rs into module
