+++ gb-cpu/src/memory/zone/mod.rs
@@ -1,0 +1,1 @@
+pub mod wram;
